update and add search-related components
